chore(rect.py): move module into `geometry/` subdirectory
